Cleaned up existing tests and merged TestGUIAudio with TestGUIHyperlink for coverage
