+++ src/main/java/gui/page/Page.java
@@ -284,4 +284,8 @@ public class Page extends StackPane implements MediaObserver {
             }
         }
     }
+
+    public Pane getMediaLayer() {
+        return mediaLayer;
+    }
 }
