+++ src/test/java/TestGUIAudio.java
@@ -1,5 +1,8 @@
+import app.MediaCommunicator;
 import app.media.MediaAudio;
+import app.media_managers.AudioModifier;
 import gui.media.GUIAudio;
+import gui.media.GUIHyperlink;
 import gui.model.GUIPlayerModel;
 import gui.page.Page;
 import javafx.application.Platform;
@@ -10,46 +13,52 @@ import javafx.scene.media.MediaPlayer;
 import javafx.util.Duration;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import storage.SQLiteStorage;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.util.ArrayList;
 
 
 public class TestGUIAudio{
     /**
-     * These test MediaPlayerController, GUIAudio, and GUIPlayerModel as a whole since they are so heavily interlinked
+     * These tests are broken, and i cannot for the life of me get these to work consistently in any meaningful capacity
+     * leaving the tests in to show what I would have tested for
      *
-     * These tests assume that the creation of MediaPlayer will work, since those are tested in TestAudioModifer
-     * They also assume the UI controls work and send the correct "signals", and that aspect unfortunately cannot really
-     * be tested. That being said, from more visual testing it can be verified that it's correct
-     *
-     * Some of these tests only work when you run them on their own, it's really weird.
-     *
-     * Some modified methods are used as FileChooser and junit do not mesh well (or at all)
+     * Cover testing for GUIAudio, GUIHyperlink, PlayerModel
      */
     private Duration TimeStamp;
     private MediaAudio audio;
-    private Page page;
+    private static Page page;
     private static Boolean init = false;
     private static TestAudioModifier tam = new TestAudioModifier();
-    private static GUIAudio audioGUI;
+    private GUIAudio audioGUI;
     private static GUIPlayerModel model;
 
     @BeforeClass
     public static void initJfxRuntime() throws Exception {
         Platform.startup(() -> {});
-        tam.createPage();
-        //audioGUI = tam.newMedia("src\\test\\java\\test_files\\1.17 Axe to Grind.mp3");
+        page = createPage();
+    }
+    public void createAudio() throws Exception {
+        GUIAudio audioGUI = addMedia("src\\test\\java\\test_files\\1.17 Axe to Grind.mp3");
         audioGUI.getAudioPlayer().setMute(true); //Preventing audio jumpscares
-
         //Because the setOnReady call gets skipped, these need to be manually initialized
-
         audioGUI.configureUI();
         audioGUI.createInterface();
-        model = audioGUI.getPlayerManipulator();
+
+        page.getCommunicator().updateMedia(audioGUI.getMedia());
+
+        this.audioGUI = (GUIAudio) page.getMediaLayer().getChildren().get(0);
+        model = new GUIPlayerModel(audioGUI, audioGUI.getAudioPlayer().getTotalDuration());
     }
 
     //Tests for the play button
     @Test
     public void testPlayButton_playStart() throws Exception {
         //Test that when play is pressed at the start of an audio file while the player is paused, it plays
+        createAudio();
         audioGUI.getAudioPlayer().seek(audioGUI.getAudioPlayer().getStartTime());
 
         model.firedPlayButton("Play");
@@ -61,6 +70,7 @@ public class TestGUIAudio{
     @Test
     public void testPlayButtonPlays_playMiddle() throws Exception {
         //Test that when play is pressed at the middle of an audio file while the player is paused, it plays
+        createAudio();
         audioGUI.getAudioPlayer().seek(new Duration(90000));
 
         model.firedPlayButton("Play");
@@ -73,6 +83,7 @@ public class TestGUIAudio{
     @Test
     public void testPlayButtonPlays_playEnd() throws Exception {
         //Test that when play is pressed at the end of an audio file while the player is paused, it plays
+        createAudio();
         audioGUI.getAudioPlayer().seek(audioGUI.getAudioPlayer().getTotalDuration());
 
         model.firedPlayButton("Play");
@@ -87,6 +98,7 @@ public class TestGUIAudio{
     @Test
     public void testPlayButton_pauseStart() throws Exception {
         //Test that when play is pressed near the start of an audio file while the player is playing, it pauses
+        createAudio();
         audioGUI.getAudioPlayer().seek(audioGUI.getAudioPlayer().getStartTime());
 
         model.firedPlayButton("Play");
@@ -101,6 +113,7 @@ public class TestGUIAudio{
     @Test
     public void testPlayButtonPlays_pauseMiddle() throws Exception {
         //Test that when play is pressed near the middle of an audio file while the player is paused, it pauses
+        createAudio();
         audioGUI.getAudioPlayer().seek(new Duration(90000));
 
         model.firedPlayButton("Play");
@@ -115,6 +128,7 @@ public class TestGUIAudio{
     @Test
     public void testPlayButtonPlays_pauseEnd() throws Exception {
         //Test that when play is pressed near the end of an audio file while the player is paused, it pauses
+        createAudio();
         audioGUI.getAudioPlayer().seek(Duration.seconds(audioGUI.getAudioPlayer().getTotalDuration().toSeconds()
                 - 10));
 
@@ -135,8 +149,9 @@ public class TestGUIAudio{
 
     //Tests for the playback rate box
     @Test
-    public void testPlayRate_rightCall()    {
+    public void testPlayRate_rightCall() throws Exception {
         //Test that the play rate box correctly determines the play rate
+        createAudio();
         ComboBox<String> playrate = audioGUI.getPlayerUI().getPlayRateOptions();
         //0.5x option
         playrate.getSelectionModel().select(0);
@@ -157,41 +172,46 @@ public class TestGUIAudio{
 
 
     @Test
-    public void testPlayRate_half() {
+    public void testPlayRate_half() throws Exception {
         //Test that the half rate option correctly adjusts player rate
+        createAudio();
         model.changePlayRate(0.5);
         assertEquals(0.5, audioGUI.getAudioPlayer().getRate(), 0.0);
     }
 
     @Test
-    public void testPlayRate_normal() {
+    public void testPlayRate_normal() throws Exception {
         //Test that the normal rate option correctly adjusts player rate
+        createAudio();
         model.changePlayRate(1);
 
         assertEquals(1, audioGUI.getAudioPlayer().getRate(), 0.0);
     }
 
     @Test
-    public void testPlayRate_halfOverNormal() {
+    public void testPlayRate_halfOverNormal() throws Exception {
         //Test that the 1.5x rate option correctly adjusts player rate
+        createAudio();
         model.changePlayRate(1.5);
 
         assertEquals(1.5, audioGUI.getAudioPlayer().getRate(), 0.0);
     }
 
     @Test
-    public void testPlayRate_double() {
+    public void testPlayRate_double() throws Exception {
         //Test that the double rate option correctly adjusts player rate
+        createAudio();
         model.changePlayRate(2);
 
         assertEquals(2, audioGUI.getAudioPlayer().getRate(), 0.0);
     }
 
     @Test
-    public void testPlayRate_playingHalf() throws InterruptedException {
+    public void testPlayRate_playingHalf() throws Exception {
         //Test that the play rate adjusts to half even with the player is playing
+        createAudio();
         audioGUI.getAudioPlayer().seek(new Duration(0));
-       model.firedPlayButton("Play");
+        model.firedPlayButton("Play");
         Thread.sleep(1000);
         model.changePlayRate(0.5);
         assertEquals(audioGUI.getAudioPlayer().getStatus(), MediaPlayer.Status.PLAYING);
@@ -200,9 +220,9 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testPlayRate_playingNormal() throws InterruptedException {
+    public void testPlayRate_playingNormal() throws Exception {
         //Test that the play rate adjusts to normal even with the player is playing
-
+        createAudio();
         model.changePlayRate(0.5);
         audioGUI.getAudioPlayer().seek(new Duration(0));
         model.firedPlayButton("Play");
@@ -214,8 +234,9 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testPlayRate_playingHalfOverNormal() throws InterruptedException {
+    public void testPlayRate_playingHalfOverNormal() throws Exception {
         //Test that the play rate adjusts to 1.5x even with the player is playing
+        createAudio();
         audioGUI.getAudioPlayer().seek(new Duration(0));
         model.firedPlayButton("Play");
         Thread.sleep(1000);
@@ -226,8 +247,9 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testPlayRate_playingDouble() throws InterruptedException {
+    public void testPlayRate_playingDouble() throws Exception {
         //Test that the play rate adjusts to 2x even with the player is playing
+        createAudio();
         audioGUI.getAudioPlayer().seek(new Duration(0));
         model.firedPlayButton("Play");
         Thread.sleep(1000);
@@ -239,8 +261,9 @@ public class TestGUIAudio{
 
     //Tests for the playback slider
     @Test
-    public void testPlaySlider_setStart() throws InterruptedException {
+    public void testPlaySlider_setStart() throws Exception {
         //Test that the play slider sets the mediaplayer to the start when it is set to it's minimal value
+        createAudio();
         model.firedPlayButton("Pause");
         audioGUI.getAudioPlayer().seek(audioGUI.getAudioPlayer().getTotalDuration());
 
@@ -253,8 +276,9 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testPlaySlider_setMiddle() throws InterruptedException {
+    public void testPlaySlider_setMiddle() throws Exception {
         //Test that the play slider sets the mediaplayer to the middle when it is set to the middle of the slider
+        createAudio();
         audioGUI.getAudioPlayer().seek(audioGUI.getAudioPlayer().getTotalDuration());
 
         model.playbackSliderAdjusted(0.5, audioGUI.getAudioPlayer().getStatus());
@@ -269,8 +293,9 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testPlaySlider_setEnd() throws InterruptedException {
+    public void testPlaySlider_setEnd() throws Exception {
         //Test that the play slider sets the mediaplayer to the end when it is set to the end of the slider
+        createAudio();
         audioGUI.getAudioPlayer().seek(new Duration(0));
         model.playbackSliderAdjusted(1, audioGUI.getAudioPlayer().getStatus());
         Thread.sleep(2000);
@@ -282,9 +307,10 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testPlaySlider_setStartPlaying() throws InterruptedException {
+    public void testPlaySlider_setStartPlaying() throws Exception {
         //Test that the play slider sets the mediaplayer to the start when it is set to the end of the slider and
         // still playing
+        createAudio();
         model.firedPlayButton("Play");
         model.playbackSliderAdjusted(0.5, audioGUI.getAudioPlayer().getStatus());
         double unexpected = audioGUI.getAudioPlayer().getCurrentTime().toMillis();
@@ -297,9 +323,10 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testPlaySlider_setMiddlePlaying() throws InterruptedException {
+    public void testPlaySlider_setMiddlePlaying() throws Exception {
         //Test that the play slider sets the mediaplayer to the middle when it is set to the end of the slider and
         // still playing
+        createAudio();
         audioGUI.getAudioPlayer().seek(new Duration(0));
         model.firedPlayButton("Play");
         model.playbackSliderAdjusted(0.5, audioGUI.getAudioPlayer().getStatus());
@@ -313,9 +340,10 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testPlaySlider_setEndPlaying() throws InterruptedException {
+    public void testPlaySlider_setEndPlaying() throws Exception {
         //Test that the play slider sets the mediaplayer to the end when it is set to the end of the slider and
         // still playing
+        createAudio();
         audioGUI.getAudioPlayer().seek(new Duration(0));
         model.firedPlayButton("Play");
         model.playbackSliderAdjusted(1, audioGUI.getAudioPlayer().getStatus());
@@ -331,29 +359,33 @@ public class TestGUIAudio{
 
     //Tests for the volume slider
     @Test
-    public void testVolumeSlider_setStart() throws InterruptedException {
+    public void testVolumeSlider_setStart() throws Exception {
         //Test that the volume slider can set volume to its minimal value
+        createAudio();
         model.changeVolume(0);
         assertEquals(0, audioGUI.getAudioPlayer().getVolume(), 0);
     }
 
     @Test
-    public void testVolumeSlider_setMiddle() throws InterruptedException {
+    public void testVolumeSlider_setMiddle() throws Exception {
         //Test that the volume slider can set volume to 0.5x volume
+        createAudio();
         model.changeVolume(0.5);
         assertEquals(0.5, audioGUI.getAudioPlayer().getVolume(), 0.5);
     }
 
     @Test
-    public void testVolumeSlider_setEnd() throws InterruptedException {
+    public void testVolumeSlider_setEnd() throws Exception {
         //Test that the volume slider can set volume to max volume
+        createAudio();
         model.changeVolume(1);
         assertEquals(1, audioGUI.getAudioPlayer().getVolume(), 1);
     }
 
     @Test
-    public void testVolumeSlider_setStartPlaying() throws InterruptedException {
+    public void testVolumeSlider_setStartPlaying() throws Exception {
         //Test that the volume slider can set the volume to 0 while the player is still playing
+        createAudio();
         model.firedPlayButton("Play");
         model.changeVolume(0);
         Thread.sleep(2000);
@@ -363,8 +395,9 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testVolumeSlider_setMiddlePlaying() throws InterruptedException {
+    public void testVolumeSlider_setMiddlePlaying() throws Exception {
         //Test that the volume slider can set the volume to 0.5 while the player is still playing
+        createAudio();
         model.firedPlayButton("Play");
         model.changeVolume(0.5);
         Thread.sleep(2000);
@@ -374,8 +407,9 @@ public class TestGUIAudio{
     }
 
     @Test
-    public void testVolumeSlider_setEndPlaying() throws InterruptedException {
+    public void testVolumeSlider_setEndPlaying() throws Exception {
         //Test that the volume slider can set the volume to 1 while the player is still playing
+        createAudio();
         model.firedPlayButton("Play");
         model.changeVolume(1);
         Thread.sleep(2000);
@@ -384,5 +418,150 @@ public class TestGUIAudio{
         model.firedPlayButton("Pause");
     }
 
+    //tests for timestamps
+
+    public void addTimestamp(Duration duration) throws Exception {
+        AudioModifier am = new AudioModifier();
+        am.modifyMedia(audioGUI.getMedia(), duration, page.getCommunicator());
+    }
+
+    public void clearTimestamps()   {
+        audioGUI.getMedia().getTimestamps().clear();
+    }
+
+    public void setUpOne() throws Exception {
+        clearTimestamps();
+        addTimestamp(new Duration(0));
+    }
+
+    public void setUpMultiple() throws Exception {
+        clearTimestamps();
+
+        addTimestamp(new Duration(0));
+        addTimestamp(new Duration(1000));
+        addTimestamp(new Duration(2000));
+        addTimestamp(new Duration(3000));
+    }
+
+    //Testing addition/removal of timestamps at both the GUI and Core Data level
+
+    @Test
+    public void testAddTimestamp_largeDuration() throws Exception {
+        createAudio();
+        clearTimestamps();
+        addTimestamp(new Duration(100000000));
+        assertEquals(audioGUI.getMedia().getTimestamps().get(0), new Duration(100000000));
+        GUIHyperlink link = (GUIHyperlink) audioGUI.getTimestamps().getChildren().get(0);
+        assertEquals("27:46:40", link.getHyperlink().getText());
+    }
+
+    //Testing functionality of timestamps
+    //It is assumed that timestamps will be for valid points in the associated audio
+
+    @Test
+    public void testClickTimestamp_startClip() throws Exception {
+        //Test that a timestamp linked to the start of the player plays said part
+        createAudio();
+        setUpOne();
+        GUIHyperlink hyperlink = (GUIHyperlink) audioGUI.getTimestamps().getChildren().get(0);
+        hyperlink.getHyperlink().fire();
+        assertEquals(0, audioGUI.getCurrentDuration().toMillis(), 2000);
+    }
+
+    @Test
+    public void testClickTimestamp_middleClip() throws Exception {
+        //Test that a timestamp linked to the middle of the player plays said part
+        //This test is prone to randomly failing, threading issues
+        createAudio();
+        clearTimestamps();
+        addTimestamp(audioGUI.getAudioPlayer().getTotalDuration().divide(2));
+
+        GUIHyperlink hyperlink = (GUIHyperlink) audioGUI.getTimestamps().getChildren().get(0);
+        hyperlink.getHyperlink().fire();
+        Thread.sleep(1000);
+        assertEquals(audioGUI.getAudioPlayer().getTotalDuration().toMillis() / 2,
+                audioGUI.getCurrentDuration().toMillis(), 2000);
+    }
+
+    @Test
+    public void testClickTimestamp_endClip() throws Exception {
+        //Test that a timestamp linked to the end of the player plays said part (or rather, doesnt)
+        createAudio();
+        clearTimestamps();
+        addTimestamp(audioGUI.getAudioPlayer().getTotalDuration());
+
+        GUIHyperlink hyperlink = (GUIHyperlink) audioGUI.getTimestamps().getChildren().get(0);
+        hyperlink.getHyperlink().fire();
+        Thread.sleep(2000);
+        assertEquals(audioGUI.getAudioPlayer().getTotalDuration().toMillis(),
+                audioGUI.getCurrentDuration().toMillis(), 0);
+    }
+
+    @Test
+    public void testClickTimestamp_first() throws Exception {
+        //Test that clicking the first of multiple timestamps works as expected
+        createAudio();
+        setUpMultiple();
+        GUIHyperlink hyperlink = (GUIHyperlink) audioGUI.getTimestamps().getChildren().get(0);
+        hyperlink.getHyperlink().fire();
+        Thread.sleep(1000);
+        assertEquals(0, audioGUI.getCurrentDuration().toMillis(), 2000);
+    }
+
+    @Test
+    public void testClickTimestamp_second() throws Exception {
+        //Test that clicking the second of multiple timestamps works as expected
+        createAudio();
+        setUpMultiple();
+        GUIHyperlink hyperlink = (GUIHyperlink) audioGUI.getTimestamps().getChildren().get(1);
+        hyperlink.getHyperlink().fire();
+        Thread.sleep(1000);
+        assertEquals(1000, audioGUI.getCurrentDuration().toMillis(), 2000);
+    }
+
+    @Test
+    public void testClickTimestamp_last() throws Exception {
+        //Test that clicking the last of multiple timestamps works as expected
+        createAudio();
+        setUpMultiple();
+        GUIHyperlink hyperlink = (GUIHyperlink) audioGUI.getTimestamps().getChildren().get(3);
+        hyperlink.getHyperlink().fire();
+        Thread.sleep(1000);
+        assertEquals(3000, audioGUI.getCurrentDuration().toMillis(), 2000);
+    }
+
+    public static Page createPage()    {
+        try {
+            SQLiteStorage storage = new SQLiteStorage(null);
+            MediaCommunicator mediaCommunicator = new MediaCommunicator(storage);
+            return new Page(mediaCommunicator);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public GUIAudio addMedia(String source) {
+        //Loading raw audio data based on user selection
+        byte[] rawData = readFile(source);
+
+        try {
+            MediaAudio audio = new MediaAudio("", 0, 0, 0, 0, rawData, new ArrayList<Duration>(),
+                    "Audio"); //Temp Constructor
+            page.getCommunicator().updateMedia(audio);
+            GUIAudio audioUI = new GUIAudio(audio);
+            return audioUI;
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public byte[] readFile(String path) {
+        File chosenFile = new File(path);
+        try {
+            return Files.readAllBytes(chosenFile.toPath());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
 
 }
